Got a headless version of the getting started example working
Got a headless version of the getting started nRF code working. Except
it isn't *quite* headless yet. It works headless, but the code still
contains some prints (which aren't necessary to the working of the code,
but are useful for debugging). Next step is to get rid of the print
statements and be completely headless. Then on to getting the PIR sensor
working wirelessly.