Significant changes to accumulator
This version of the accumulator is significantly more stable than
previous versions. Still need to make sure the speaker is low when
supposed to be low and high when supposed to be high.

Also the circuit seems to have something wrong with it - the PIR works
fine, but using the door switch causes it to freeze, even though there
is no difference in the code (both are on pin 3, low signal and are
treated identically - in fact, the Arduino can't even tell the
difference).