First stable version!
Hurray!

Every node can arm and disarm, sense motion, and send their signals
reliably!

Known issues still to be tracked down:
-The arm/disarm node seems to freeze eventually. A hack is to reset it
every so often, and I will do this. But I need to figure out why it is
freezing so I can get at the root of the matter.
-The arm/disarm node needs to be updated with a new RFID reader which
can trigger an interrupt based on the presence of a card to scan and
which will then communicate through the RX pin on the Arduino.
-The accumulator node freezes when its magnetic switch is open.